extracted more to utils.ts
